Additional post metadata - type
Posts can have a 'type' attribute. I want my blog to have different
categories of posts, that are exactly like posts in all respects but are
styled differently ( linkblog entries, photo galleries etc. )
Adding this as metadata makes it easy to wrap posts in a container element to
attach styles to.